chore(tailwind): remove uneccesseary folder nesting
