+++ packages/orbit-tailwind-preset/src/__tests__/helpers.test.ts
@@ -1,4 +1,4 @@
-import { getComponentLevelTokens } from "../presets/foundation/helpers";
+import getComponentLevelTokens from "../getComponentLevelTokens";
 
 describe("orbit-tailwind-preset", () => {
   it("should get component level token", () => {
