+++ packages/orbit-tailwind-preset/src/foundation/cssVarsFoundation.ts
