+++ packages/orbit-tailwind-preset/src/utils/firstToUpper.ts
@@ -1,0 +1,3 @@
+const firstToUpper = (str: string) => str.charAt(0).toUpperCase() + str.slice(1);
+
+export default firstToUpper;
