+++ packages/orbit-tailwind-preset/src/utils/kebabCase.ts
@@ -1,0 +1,3 @@
+const kebabCase = (str: string) => str.replace(/([a-z])([A-Z])/g, "$1-$2").toLowerCase();
+
+export default kebabCase;
