+++ packages/orbit-tailwind-preset/tailwind.config.ts
@@ -2,6 +2,6 @@ import orbitComponentsPreset from "./dist";
 
 export default {
   // include only for testing purposes
-  content: ["./src/presets/__fixtures__/*.tsx"],
+  content: ["./src/__fixtures__/*.tsx"],
   presets: [orbitComponentsPreset()],
 };
